move test-independent execution stuff outside test command package
