+++ protostar/commands/test/cheatcodes/mock_call_cheatcode.py
@@ -4,8 +4,8 @@ from typing import Any, Callable, Dict, List, Optional, Union
 from starkware.starknet.public.abi import get_selector_from_name
 from starkware.starknet.public.abi import AbiType
 
-from protostar.commands.test.starkware.cheatcode import Cheatcode
-from protostar.commands.test.starkware.types import AddressType
+from protostar.starknet.cheatcode import Cheatcode
+from protostar.starknet.types import AddressType
 from protostar.commands.test.test_environment_exceptions import CheatcodeException
 from protostar.utils.data_transformer_facade import DataTransformerFacade
