+++ protostar/commands/test/cheatcodes/warp_cheatcode.py
@@ -1,7 +1,7 @@
 from typing import Any, Callable, Optional
 
-from protostar.commands.test.starkware.cheatcode import Cheatcode
-from protostar.commands.test.starkware.types import AddressType
+from protostar.starknet.cheatcode import Cheatcode
+from protostar.starknet.types import AddressType
 
 
 class WarpCheatcode(Cheatcode):
