--- protostar/commands/test/starkware/__init__.py
@@ -1,9 +1,0 @@
-from protostar.commands.test.starkware.cheatable_syscall_handler import (
-    CheatableHintsWhitelist,
-    CheatableSysCallHandler,
-    CheatableSysCallHandlerException,
-)
-from protostar.commands.test.starkware.execution_resources_summary import (
-    ExecutionResourcesSummary,
-)
-from protostar.commands.test.starkware.forkable_starknet import ForkableStarknet
