+++ protostar/starknet/cheatable_state.py
@@ -21,10 +21,10 @@ from starkware.starknet.testing.state import StarknetState
 from starkware.storage.dict_storage import DictStorage
 from starkware.storage.storage import FactFetchingContext
 
-from protostar.commands.test.starkware.cheatable_execute_entry_point import (
+from protostar.starknet.cheatable_execute_entry_point import (
     CheatableExecuteEntryPoint,
 )
-from protostar.commands.test.starkware.types import (
+from protostar.starknet.types import (
     AddressType,
     ClassHashType,
     SelectorType,
